Automatically setup file permissions
This introduces an extra entrypoint.sh script that is called first. It
will do a chown on the entire osh directory, ensuring that any volumes
mounted from the host system will have the right owner and will be
writable by the osh process.

This requires running the entrypoint as root instead of having docker
drop to the osh user, but entrypoint.sh makes sure to drop to the osh
user before executing launch.sh.