Make image/tag names consistent in the README
Docker is a bit inconsistent in its naming, since it seems to call the
argument to `docker build --tag` a tag, but also just the "latest" part
of "foo:latest". This inconsistency was reflected in the README where
sometimes a `[tag]` placeholder was used and sometimes
`<repository>:<tag>` to refer to the same thing.

This makes things consistent by using `[image]` everywhere to mean an
identifier to a (tagged) image.