Small improvements (#3)
* Automatically setup file permissions

This introduces an extra entrypoint.sh script that is called first. It
will do a chown on the entire osh directory, ensuring that any volumes
mounted from the host system will have the right owner and will be
writable by the osh process.

This requires running the entrypoint as root instead of having docker
drop to the osh user, but entrypoint.sh makes sure to drop to the osh
user before executing launch.sh.

* Recommend using tags instead of branches when possible

* Fix typo in docker build command

* Remove mention of ports for docker-compose in README

There are already reasonable defaults for the port mapping in the
docker-compose file, so no need for the README to say these need to be
added.

* Do a shallow clone in the docker build

No need to get all git history, since the clone will be thrown away
after the build anyway.

* Fix ports in example docker run commands

This fixes the HTTPS port and also adds the HTTP port, as listed in the
Dockerfile EXPOSE command.

* Remove duplicate section from README

* Make image/tag names consistent in the README

Docker is a bit inconsistent in its naming, since it seems to call the
argument to `docker build --tag` a tag, but also just the "latest" part
of "foo:latest". This inconsistency was reflected in the README where
sometimes a `[tag]` placeholder was used and sometimes
`<repository>:<tag>` to refer to the same thing.

This makes things consistent by using `[image]` everywhere to mean an
identifier to a (tagged) image.

* Default to builing osh-node-dev-template in docker-compose

This ensure that the file in the repository actually produces
a meaningful result without needing to modify anything.

Additionally, update the docker-compose documentation to show three
different image retrieval strategies for docker compose, and some more
compose documentation improvements.

* Add a TL;DR to the README